Thank you very much for providing me with an interesting challenge which I am enjoying. Also, thanks for the quick response to my queries.
Please review my attempt hosted at http://sympytest.appspot.com/

My source code is at https://github.com/kshitijaranke/sympy-live

Features suggested in the task:

1) Adding a feed of recent searches

--> I have added a box to the right of the page which displays the ten most recent searches.

2) Privacy mode

--> I have added an option in the toolbar to keep searches private.

3) Executing recent searches

--> At present, the user can copy-paste from the recent searches into the terminal to execute the query. I am not sure this is the best approach. Any ideas you have are welcome.

Please help me if I have made any errors as I am not familiar with git and ext-js.

I have modified the following parts of the code:

1) shell.py

I have added a Searches class which is a blueprint for storing a query and it's corresponding timestamp to a datastore. In class FrontPageHandler, the 10 most recent queries are searched and added to results. In EvaluateHandler, I have added a variable(query_forbidden) which reads the value of the privacy option and stores the query only when allowed.

2) templates/shell.html

I have added a div(recent_searches) which prints all the recent queries entered.

3) static/live-core.js

I have added a variable(recordTypes) to represent the privacy checkbox options with default value 'no'. In the toolbar, I have added another select(recordEl). This parameter is passed with the key 'forbidden'.

Looking forward to your comments.